tests/requirements.txt
pytest
pytest-mock
boto3
Flask-lambda
requests